lambda/reviews/requirements.txt
boto3
PyJWT==2.8.0
requests